Move field parsers to own subpackage
