+++ includes/class-plugin.php
@@ -10,7 +10,7 @@
 
 namespace WP_Business_Reviews\Includes;
 
-use WP_Business_Reviews\Includes\Field\Field_Parser;
+use WP_Business_Reviews\Includes\Field\Parser\Plugin_Settings_Field_Parser;
 use WP_Business_Reviews\Includes\Field\Field_Repository;
 use WP_Business_Reviews\Includes\Serializer\Option_Serializer;
 use WP_Business_Reviews\Includes\Deserializer\Option_Deserializer;
@@ -104,11 +104,11 @@ final class Plugin {
 			$field_factory = new Field_Factory();
 
 			// Register settings.
-			$plugin_settings_config = new Config( WPBR_PLUGIN_DIR . 'configs/config-plugin-settings.php' );
-			$settings_field_parser  = new Field_Parser( $option_deserializer, $field_factory );
-			$plugin_settings        = new Plugin_Settings(
+			$plugin_settings_config       = new Config( WPBR_PLUGIN_DIR . 'configs/config-plugin-settings.php' );
+			$plugin_settings_field_parser = new Plugin_Settings_Field_Parser( $option_deserializer, $field_factory );
+			$plugin_settings              = new Plugin_Settings(
 				$plugin_settings_config,
-				$settings_field_parser,
+				$plugin_settings_field_parser,
 				$platform_manager->get_active_platforms(),
 				$platform_manager->get_connected_platforms()
 			);
